+++ src/plugins/plausible.plugin.ts
@@ -1,4 +1,5 @@
 import { config } from '@/config';
+
 import Plausible from 'plausible-tracker';
 import type { App } from 'vue';
 
@@ -7,6 +8,6 @@ export const plausible = {
     const plausible = Plausible(config.plausible);
     plausible.enableAutoPageviews();
 
-    app.config.globalProperties.$plausible = plausible;
+    app.provide('plausible', plausible);
   },
 };